# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GHF Inventarios API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /marcas:
    get:
      operationId: listMarcas
      tags: [marcas]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Marca" }
    post:
      operationId: createMarca
      tags: [marcas]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MarcaInput" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Marca" }
  /marcas/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getMarca
      tags: [marcas]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Marca" }
    put:
      operationId: updateMarca
      tags: [marcas]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MarcaInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Marca" }
    delete:
      operationId: deleteMarca
      tags: [marcas]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  # ------------------ LOCALES ------------------
  /locales:
    get:
      operationId: listLocales
      tags: [locales]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Local" }
    post:
      operationId: createLocal
      tags: [locales]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LocalInput" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Local" }
  /locales/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getLocal
      tags: [locales]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Local" }
    put:
      operationId: updateLocal
      tags: [locales]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LocalInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Local" }
    delete:
      operationId: deleteLocal
      tags: [locales]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  # ------------------ CATEGORIAS ------------------
...
  /pedidos:
    get:
      operationId: listPedidos
      tags: [pedidos]
      parameters:
        - in: query
          name: estado
          schema: { type: string }
        - in: query
          name: localId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Pedido" }
    post:
      operationId: createPedido
      tags: [pedidos]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PedidoInput" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pedido" }
  /pedidos/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getPedido
      tags: [pedidos]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pedido" }
    delete:
      operationId: anularPedido
      tags: [pedidos]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /pedidos/{id}/facturar:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    post:
      operationId: facturarPedido
      tags: [pedidos]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Factura" }

  # ------------------ FACTURAS ------------------
...
components:
  schemas:
...
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    Marca:
      type: object
      properties:
        id: { type: integer }
        nombre: { type: string }
        slug: { type: string, nullable: true }
        activo: { type: boolean }
        localesCount: { type: integer }
      required: [id, nombre, activo, localesCount]
    MarcaInput:
      type: object
      properties:
        nombre: { type: string }
        slug: { type: string, nullable: true }
        activo: { type: boolean }
      required: [nombre]
    Local:
      type: object
      properties:
        id: { type: integer }
        marcaId: { type: integer, nullable: true }
        sociedadId: { type: integer, nullable: true }
        marcaNombre: { type: string, nullable: true }
        sociedadNombre: { type: string, nullable: true }
        nombre: { type: string }
        codigo: { type: string, nullable: true }
        correo: { type: string, nullable: true }
        telefono: { type: string, nullable: true }
        direccion: { type: string, nullable: true }
        activo: { type: boolean }
      required: [id, nombre, activo]
    LocalInput:
      type: object
      properties:
        marcaId: { type: integer, nullable: true }
        sociedadId: { type: integer, nullable: true }
        nombre: { type: string }
        codigo: { type: string, nullable: true }
        correo: { type: string, nullable: true }
        telefono: { type: string, nullable: true }
        direccion: { type: string, nullable: true }
        activo: { type: boolean }
      required: [nombre]
...
    PedidoDetalle:
      type: object
      properties:
        id: { type: integer }
        productoId: { type: integer }
        productoNombre: { type: string }
        cantidad: { type: integer }
        precioUnitario: { type: number }
        subtotal: { type: number }
      required: [id, productoId, productoNombre, cantidad, precioUnitario, subtotal]
    PedidoDetalleInput:
      type: object
      properties:
        productoId: { type: integer }
        cantidad: { type: integer }
        precioUnitario: { type: number }
      required: [productoId, cantidad, precioUnitario]
    Pedido:
      type: object
      properties:
        id: { type: integer }
        localId: { type: integer, nullable: true }
        localNombre: { type: string, nullable: true }
        clienteNombre: { type: string, nullable: true }
        clienteTelefono: { type: string, nullable: true }
        clienteEmail: { type: string, nullable: true }
        observaciones: { type: string, nullable: true }
        subtotal: { type: number }
        impuesto: { type: number }
        total: { type: number }
        estado: { type: string }
        createdAt: { type: string }
        detalles:
          type: array
          items: { $ref: "#/components/schemas/PedidoDetalle" }
      required: [id, subtotal, impuesto, total, estado, createdAt, detalles]
    PedidoInput:
      type: object
      properties:
        localId: { type: integer, nullable: true }
        clienteNombre: { type: string, nullable: true }
        clienteTelefono: { type: string, nullable: true }
        clienteEmail: { type: string, nullable: true }
        observaciones: { type: string, nullable: true }
        impuestoPct: { type: number, default: 0 }
        detalles:
          type: array
          items: { $ref: "#/components/schemas/PedidoDetalleInput" }
      required: [detalles]
    FacturaDetalle:
      type: object
      properties:
        id: { type: integer }
        productoId: { type: integer, nullable: true }
        descripcion: { type: string }
        cantidad: { type: integer }
        precioUnitario: { type: number }
        subtotal: { type: number }
      required: [id, descripcion, cantidad, precioUnitario, subtotal]
    Factura:
      type: object
      properties:
        id: { type: integer }
        pedidoId: { type: integer, nullable: true }
        localId: { type: integer, nullable: true }
        localNombre: { type: string, nullable: true }
        numeroFactura: { type: string }
        clienteNombre: { type: string, nullable: true }
        clienteTelefono: { type: string, nullable: true }
        clienteEmail: { type: string, nullable: true }
        fecha: { type: string }
        subtotal: { type: number }
        impuesto: { type: number }
        total: { type: number }
        estado: { type: string }
        observaciones: { type: string, nullable: true }
        detalles:
          type: array
          items: { $ref: "#/components/schemas/FacturaDetalle" }
      required: [id, numeroFactura, fecha, subtotal, impuesto, total, estado, detalles]
...