# lib/api-zod/src/generated/types/marca.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GHF Inventarios API
 * OpenAPI spec version: 0.1.0
 */

export interface Marca {
  id: number;
  nombre: string;
  slug?: string | null;
  activo: boolean;
  localesCount: number;
}